+++ source/sal_mbr.h
@@ -2,4 +2,4 @@
 
 #include "sal.h"
 
-int read_usb_partition_from_mbr(FSSALAttachDeviceArg *attach_arg, u32* out_offset, u32* out_size, u32* out_umsBlkDevID);
+int read_usb_partition_from_mbr(FSSALAttachDeviceArg *attach_arg, u32* out_offset, u32* out_size, u8* out_umsBlkDevID);
